+++ package.json
@@ -24,6 +24,7 @@
     "@mdx-js/loader": "^3.1.0",
     "@mdx-js/react": "^3.1.0",
     "@next/mdx": "^15.2.4",
+    "@rehype-pretty/transformers": "^0.13.2",
     "@types/mdx": "^2.0.13",
     "@types/node": "22.14.0",
     "@types/react": "19.1.0",
@@ -81,6 +82,7 @@
     "@storybook/testing-library": "0.2.2",
     "@svgr/webpack": "^8.1.0",
     "@tailwindcss/postcss": "^4.1.3",
+    "html-react-parser": "^5.2.3",
     "isomorphic-git": "^1.30.1",
     "postcss": "^8.5.3",
     "postcss-selector-not": "^8.0.1",
